refactor: 모든 도메인 패키지 하위에 entity 패키지 생성
