+++ BE/src/main/java/ifive/idrop/domain/auth/filter/VerifyUserFilter.java
@@ -5,7 +5,7 @@ import ifive.idrop.domain.auth.dto.LoginRequest;
 import ifive.idrop.common.enums.Role;
 import ifive.idrop.common.exception.BusinessException;
 import ifive.idrop.common.dto.ErrorResponse;
-import ifive.idrop.domain.auth.AuthenticateUser;
+import ifive.idrop.domain.auth.entity.AuthenticateUser;
 import ifive.idrop.domain.user.service.UserService;
 import jakarta.servlet.*;
 import jakarta.servlet.http.HttpServletRequest;
