+++ BE/src/main/java/ifive/idrop/domain/auth/resolver/LoginUsersArgumentResolver.java
@@ -4,10 +4,10 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import ifive.idrop.domain.auth.filter.VerifyUserFilter;
 import ifive.idrop.domain.auth.utils.JwtProvider;
-import ifive.idrop.domain.user.User;
+import ifive.idrop.domain.user.entity.User;
 import ifive.idrop.common.exception.BusinessException;
 import ifive.idrop.common.exception.ErrorCode;
-import ifive.idrop.domain.auth.AuthenticateUser;
+import ifive.idrop.domain.auth.entity.AuthenticateUser;
 import ifive.idrop.domain.user.repository.UserRepository;
 import io.jsonwebtoken.Claims;
 import jakarta.servlet.http.HttpServletRequest;
