+++ BE/src/main/java/ifive/idrop/domain/driver/dto/DriverTodayRemainingPickUpResponse.java
@@ -1,8 +1,8 @@
 package ifive.idrop.domain.driver.dto;
 
 import com.fasterxml.jackson.annotation.JsonUnwrapped;
-import ifive.idrop.domain.subscription.Subscription;
-import ifive.idrop.domain.pickup.PickUpLocation;
+import ifive.idrop.domain.subscription.entity.Subscription;
+import ifive.idrop.domain.pickup.entity.PickUpLocation;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
