+++ BE/src/main/java/ifive/idrop/domain/user/entity/User.java
@@ -1,12 +1,12 @@
-package ifive.idrop.domain.user;
+package ifive.idrop.domain.user.entity;
 
-import ifive.idrop.domain.auth.Auth;
+import ifive.idrop.domain.auth.entity.Auth;
 import ifive.idrop.domain.auth.dto.LoginRequest;
 import ifive.idrop.common.enums.Gender;
 import ifive.idrop.common.exception.BusinessException;
 import ifive.idrop.common.exception.ErrorCode;
 import ifive.idrop.domain.driver.entity.Driver;
-import ifive.idrop.domain.parent.Parent;
+import ifive.idrop.domain.parent.entity.Parent;
 import ifive.idrop.domain.user.dto.SignUpRequest;
 import jakarta.persistence.*;
 import lombok.Getter;
