+++ BE/src/main/java/ifive/idrop/websocket/location/LocationWebSocketHandler.java
@@ -4,15 +4,15 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import ifive.idrop.domain.driver.entity.Driver;
 import ifive.idrop.common.exception.BusinessException;
 import ifive.idrop.common.exception.ErrorCode;
-import ifive.idrop.domain.auth.AuthenticateUser;
+import ifive.idrop.domain.auth.entity.AuthenticateUser;
 import ifive.idrop.domain.auth.filter.VerifyUserFilter;
 import ifive.idrop.domain.auth.utils.JwtProvider;
-import ifive.idrop.domain.parent.Parent;
-import ifive.idrop.domain.pickup.PickUpHistory;
-import ifive.idrop.domain.pickup.PickUpLocation;
+import ifive.idrop.domain.parent.entity.Parent;
+import ifive.idrop.domain.pickup.entity.PickUpHistory;
+import ifive.idrop.domain.pickup.entity.PickUpLocation;
 import ifive.idrop.domain.pickup.repository.PickUpRepository;
 import ifive.idrop.domain.user.repository.UserRepository;
-import ifive.idrop.domain.user.User;
+import ifive.idrop.domain.user.entity.User;
 import ifive.idrop.websocket.CustomObjectMapper;
 import ifive.idrop.websocket.location.dto.ChildGeoLocation;
 import ifive.idrop.websocket.location.dto.CurrentPickUp;
